rsz: Promote Rebuffer2 to Rebuffer
